Remove views, forms, mixins and corresponding tests from projects app
