--- rdmo/projects/views/__init__.py
@@ -1,31 +1,0 @@
-from .integration import IntegrationCreateView, IntegrationDeleteView, IntegrationUpdateView, IntegrationWebhookView
-from .invite import InviteDeleteView
-from .issue import IssueDetailView, IssueSendView, IssueUpdateView
-from .membership import MembershipCreateView, MembershipDeleteView, MembershipUpdateView
-from .project import (
-    OldProjectDetailView,
-    ProjectCancelView,
-    ProjectDeleteView,
-    ProjectDetailView,
-    ProjectErrorView,
-    ProjectExportView,
-    ProjectInterviewView,
-    ProjectJoinView,
-    ProjectLeaveView,
-    ProjectsView,
-)
-from .project_answers import ProjectAnswersExportView, ProjectAnswersView
-from .project_copy import ProjectCopyView
-from .project_create import ProjectCreateImportView, ProjectCreateView
-from .project_update import (
-    ProjectUpdateCatalogView,
-    ProjectUpdateImportView,
-    ProjectUpdateInformationView,
-    ProjectUpdateParentView,
-    ProjectUpdateTasksView,
-    ProjectUpdateView,
-    ProjectUpdateViewsView,
-    ProjectUpdateVisibilityView,
-)
-from .project_view import ProjectViewExportView, ProjectViewView
-from .snapshot import SnapshotCreateView, SnapshotExportView, SnapshotRollbackView, SnapshotUpdateView
